Enforce exporting resources in its storage version
After cluster-api is upgraded, ownerReferences of some resources are still old group
versions. For example, in MachineDeployment's ownerReferences the Cluster's apiVersion
is still cluster.x-k8s.io/v1alpha3. It causes restore error in clusterctl move or
backup/restore if it is not the storage version v1beta1.

This patch enforces all resources are exported in its storage version.